+++ src/aliases/utility/check_wholist.lua
@@ -1,3 +1,0 @@
-enableTrigger("Parse wholist")
-send("who b")
-tempTimer(10, [[disableTrigger'Parse wholist']])