+++ src/scripts/clearlastupdate.lua
@@ -1,4 +1,0 @@
-function mapper.clearlastupdate()
-	-- reset mapper.pdb_lastupdate on the prompt
-	enableTrigger("Mudlet Mapper prompt trigger")
-end