+++ src/triggers/game_specific/gomud/login.lua
@@ -1,2 +1,0 @@
-raiseEvent("mapper logged in", "gomud")
-mapper.echo("We're connected to Willowdale.")